Attempt to Correct 2nd Error Thrown for @knathan54
Per the comment from knathan, I have moved the call to `loadBrowserList` outside of the `try...catch` block, as when there is no settings file, the browsers were not getting loaded into the combo. I still have a minor interface goof that can be fixed after this PR is merged that keeps the 'OK' button from being enabled when the data in the fields on the right side of the dialog are changed.

After reviewing the `latest-changes-2021-11-01.txt` file attached to his comment, and without any information as to exactly when these errors are being thrown, I did the best I could in deciding what changes need to be made. In an attempt to replicate what @knathan must have been getting, I deleted the `~/.NTA/` folder and all of its subfolders in hopes of getting the same error.

After doing so and running the project, The main window opens as expected. I opened the Options dialog and it opened fine. However, when I clicked the OK button, I received an error, which I traced back to the browsers list not being loaded, thereby there being no value for the line `props.setSystemProperty("default.browser", webBrowsersList.getSelectedItem().toString());` Due to there being no browsers loaded, there is no selected item, which resulted in an NPE. I have fixed this issue by removing the call to `loadBrowserList` outside of the `try...catch` block in the `loadSettings` method.

At this point, with and without files on the local system, I am receiving no exceptions at all.

@knathan54, could you pull these changes down and test again. If you have any more exceptions be thrown, please provide me step-by-step instructions on what you did leading up to the exception. Thank you.

-SC